--- src/pages/movimentacao/MovimentacaoButtons.tsx
@@ -1,13 +1,0 @@
-import { Button, Flex } from '@chakra-ui/react';
-  
-export function MovimentacaoCancelButton() {
-  return (
-    <Flex gap="60px" mt="720px" alignSelf="flex-start" marginLeft="200px">
-      <Button background="#212121" borderRadius="50px">
-        Cancelar
-      </Button>
-      <Button borderRadius="50px">Confirmar</Button>
-    </Flex>
-  );
-}
-
